Bug fix, now only allow guests to vote to skip once per song. New feature, allow for the host to go back to previous song.

+++ music_controller/spotify/views.py
@@ -116,6 +116,7 @@ class PauseSong(APIView):
     def put(self, response, format=None):
         room_code = self.request.session.get('room_code')
         room = Room.objects.filter(code=room_code)[0]
+
         if self.request.session.session_key == room.host or room.guest_can_pause:
             pause_song(room.host)
             return Response({}, status=status.HTTP_204_NO_CONTENT)
@@ -126,6 +127,7 @@ class PlaySong(APIView):
     def put(self, response, format=None):
         room_code = self.request.session.get('room_code')
         room = Room.objects.filter(code=room_code)[0]
+
         if self.request.session.session_key == room.host or room.guest_can_pause:
             play_song(room.host)
             return Response({}, status=status.HTTP_204_NO_CONTENT)
@@ -139,7 +141,8 @@ class SkipSong(APIView):
         votes = Vote.objects.filter(room=room, song_id=room.current_song)
         votes_to_skip = room.votes_to_skip
 
-        if self.request.session.session_key == room.host or len(votes) + 1 >= votes_to_skip:
+        if self.request.session.session_key == room.host or (len(votes) + 1 >= votes_to_skip and len(Vote.objects.filter(user=self.request.session.session_key, 
+                        room=room, song_id=room.current_song)) == 0):
             votes.delete()
             skip_song(room.host)
         else:
@@ -148,3 +151,13 @@ class SkipSong(APIView):
             vote.save()
 
         return Response({}, status.HTTP_204_NO_CONTENT)
+    
+class PreviousSong(APIView):
+    def post(self, request, format=None):
+        room_code = self.request.session.get('room_code')
+        room = Room.objects.filter(code=room_code)[0]
+
+        if self.request.session.session_key == room.host:
+            previous_song(room.host)
+
+        return Response({}, status.HTTP_403_FORBIDDEN)
